# ChiaraAndroid/app/src/main/openapi/chiara.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Chiara",
    "version": "1.0"
  },
  "paths": {
    "/api/Album": {
      "get": {
        "tags": [
          "Album"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlbumResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Album/{id}": {
      "get": {
        "tags": [
          "Album"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/File/{key}": {
      "get": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hls/{file}": {
      "get": {
        "tags": [
          "Hls"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Hls/start/{episodeId}": {
      "post": {
        "tags": [
          "Hls"
        ],
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPlayingResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPlayingResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPlayingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Hls/stop": {
      "post": {
        "tags": [
          "Hls"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MediaRepository": {
      "get": {
        "tags": [
          "MediaRepository"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaRepositoryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaRepositoryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaRepositoryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MediaRepository"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaRepositoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaRepositoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaRepositoryRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/MediaRepository/{id}": {
      "get": {
        "tags": [
          "MediaRepository"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MediaRepository"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRepositoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Season": {
      "get": {
        "tags": [
          "Season"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShowSeasonResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShowSeasonResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShowSeasonResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Season/{id}": {
      "get": {
        "tags": [
          "Season"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSeasonResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSeasonResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowSeasonResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Song": {
      "get": {
        "tags": [
          "Song"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Song/{id}": {
      "get": {
        "tags": [
          "Song"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlbumResponse": {
        "required": [
          "arist",
          "coverImageUrl",
          "id",
          "songs",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "arist": {
            "minLength": 1,
            "type": "string"
          },
          "coverImageUrl": {
            "minLength": 1,
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "CreateMediaRepositoryRequest": {
        "required": [
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "path": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "EpisodeResponse": {
        "required": [
          "episodeNumber",
          "id",
          "showSeasonId",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "episodeNumber": {
            "minLength": 1,
            "type": "string"
          },
          "showSeasonId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "MediaRepositoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "albumResponses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlbumResponse"
            },
            "nullable": true
          },
          "showSeasonResponses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShowSeasonResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ShowSeasonResponse": {
        "required": [
          "episodes",
          "id",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeResponse"
            }
          }
        },
        "additionalProperties": false
      },
      "SongResponse": {
        "required": [
          "arist",
          "coverImageUrl",
          "id",
          "title",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "arist": {
            "minLength": 1,
            "type": "string"
          },
          "coverImageUrl": {
            "minLength": 1,
            "type": "string"
          },
          "url": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "VideoPlayingResponse": {
        "required": [
          "playList"
        ],
        "type": "object",
        "properties": {
          "playList": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}